Initial Setup For Multi Step Registration
- Implement children routing
- Create user info form validation type
- Move code from Register.tsx into CreateAccount as Register will serve as parent route component
