+++ src/main.tsx
@@ -12,6 +12,8 @@ import ProtectedRoute from "./pages/ProtectedRoute.tsx";
 import { UserProvider } from "./contexts/UserContext.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ProjectsProvider } from "./contexts/ProjectsContext.tsx";
+import CreateUserAccount from "./pages/Auth/CreateAccount/CreateAccount.tsx";
+import CreateUserInfo from "./pages/Auth/CreateUserInfo/CreateUserInfo.tsx";
 
 const queryClient = new QueryClient();
 
@@ -46,7 +48,15 @@ const router = createBrowserRouter([
 		),
 	},
 	{ path: "/login", element: <Login /> },
-	{ path: "/register", element: <Register /> },
+	// { path: "/register", element: <Register /> },
+	{
+		path: "/register",
+		element: <Register />,
+		children: [
+			{ path: "", element: <CreateUserAccount /> },
+			{ path: "userinfo", element: <CreateUserInfo /> },
+		],
+	},
 ]);
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
