+++ src/pages/Auth/CreateUserInfo/CreateUserInfo.tsx
@@ -1,0 +1,9 @@
+const CreateUserInfo = () => {
+	return (
+		<>
+			<h1>create user info</h1>
+		</>
+	);
+};
+
+export default CreateUserInfo;
