+++ src/pages/Auth/model/index.tsx
@@ -34,3 +34,15 @@ export const registerSchema = z
 	);
 
 export type TRegisterFormInputs = z.infer<typeof registerSchema>;
+
+export const userInfoSchema = z.object({
+	username: z
+		.string()
+		.min(1, { message: "Can't be empty" })
+		.toLowerCase()
+		.trim(),
+	firstName: z.string().min(1, { message: "Can't be empty" }).trim(),
+	lastName: z.string().trim().optional(),
+});
+
+export type TUserInfoInputs = z.infer<typeof userInfoSchema>;
